move globals up one level
